semitone and scalar distance calculation
The distance of (a, b) was calculated as (a - b) but should be
calculated as (b - a). This has been fixed in both the scalar and
semitone distance methods. Since these methods are used in the invert
method, the order of arguments has been adjusted there as well.

The scalar distance calculation was both overly complex and wrong. An
approach similar to the one used to calculate semitone distance is
used instead.